chore: Refactor to move all artifacts to lib folder
