+++ .github/workflows/release.yml
@@ -16,3 +16,5 @@ jobs:
   release:
     uses: cloudscape-design/actions/.github/workflows/release.yml@main
     secrets: inherit
+    with:
+      publish-packages: lib
