+++ .gitignore
@@ -3,8 +3,4 @@ coverage
 node_modules
 
 # generated typescript artifacts
-dist
-
-# generated re-exports for public files
-/*.js
-/*.d.ts
+lib
