+++ test/utils/stop-chromedriver.ts
@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { shutdownWebdriver } from '../../dist/chrome-launcher';
+import { shutdownWebdriver } from '../../lib/chrome-launcher';
 
 export default shutdownWebdriver;
